+++ packages/orq-rc/src/models/operations/createeval.ts
@@ -378,6 +378,7 @@ export type Llm2 = {
   prompt: string;
   categories?: Array<string> | undefined;
   categoricalLabels?: Array<LLMCategoricalLabels> | undefined;
+  datasetId?: string | undefined;
   /**
    * Entity storage path.
    *
@@ -475,6 +476,7 @@ export type Llm1 = {
   prompt: string;
   categories?: Array<string> | undefined;
   categoricalLabels?: Array<CreateEvalLLMCategoricalLabels> | undefined;
+  datasetId?: string | undefined;
   /**
    * Entity storage path.
    *
@@ -1237,6 +1239,7 @@ export type CreateEvalLlm2 = {
   prompt: string;
   categories?: Array<string> | undefined;
   categoricalLabels?: Array<CreateEvalLLMEvalsCategoricalLabels> | undefined;
+  datasetId?: string | undefined;
   key: string;
   mode: "jury";
   jury: CreateEvalLLMJury;
@@ -1315,6 +1318,7 @@ export type CreateEvalLlm1 = {
   categoricalLabels?:
     | Array<CreateEvalLLMEvalsResponseCategoricalLabels>
     | undefined;
+  datasetId?: string | undefined;
   key: string;
   mode: "single";
   model: string;
@@ -2288,6 +2292,7 @@ export type Llm2$Outbound = {
   prompt: string;
   categories?: Array<string> | undefined;
   categorical_labels?: Array<LLMCategoricalLabels$Outbound> | undefined;
+  dataset_id?: string | undefined;
   path: string;
   description: string;
   key: string;
@@ -2319,6 +2324,7 @@ export const Llm2$outboundSchema: z.ZodType<Llm2$Outbound, z.ZodTypeDef, Llm2> =
     categoricalLabels: z.array(
       z.lazy(() => LLMCategoricalLabels$outboundSchema),
     ).optional(),
+    datasetId: z.string().optional(),
     path: z.string(),
     description: z.string().default(""),
     key: z.string(),
@@ -2329,6 +2335,7 @@ export const Llm2$outboundSchema: z.ZodType<Llm2$Outbound, z.ZodTypeDef, Llm2> =
       guardrailConfig: "guardrail_config",
       outputType: "output_type",
       categoricalLabels: "categorical_labels",
+      datasetId: "dataset_id",
     });
   });
 
@@ -2530,6 +2537,7 @@ export type Llm1$Outbound = {
   categorical_labels?:
     | Array<CreateEvalLLMCategoricalLabels$Outbound>
     | undefined;
+  dataset_id?: string | undefined;
   path: string;
   description: string;
   key: string;
@@ -2561,6 +2569,7 @@ export const Llm1$outboundSchema: z.ZodType<Llm1$Outbound, z.ZodTypeDef, Llm1> =
     categoricalLabels: z.array(
       z.lazy(() => CreateEvalLLMCategoricalLabels$outboundSchema),
     ).optional(),
+    datasetId: z.string().optional(),
     path: z.string(),
     description: z.string().default(""),
     key: z.string(),
@@ -2571,6 +2580,7 @@ export const Llm1$outboundSchema: z.ZodType<Llm1$Outbound, z.ZodTypeDef, Llm1> =
       guardrailConfig: "guardrail_config",
       outputType: "output_type",
       categoricalLabels: "categorical_labels",
+      datasetId: "dataset_id",
     });
   });
 
@@ -2776,8 +2786,8 @@ export const Typescript$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-05-18T05:04:24.675Z"),
-  updated: z.string().default("2026-05-18T05:04:24.675Z"),
+  created: z.string().default("2026-05-18T14:04:16.993Z"),
+  updated: z.string().default("2026-05-18T14:04:16.993Z"),
   guardrail_config: z.nullable(
     z.union([
       z.lazy(() =>
@@ -2965,8 +2975,8 @@ export const Ragas$inboundSchema: z.ZodType<Ragas, z.ZodTypeDef, unknown> = z
   .object({
     _id: z.string(),
     description: z.string(),
-    created: z.string().default("2026-05-18T05:04:24.675Z"),
-    updated: z.string().default("2026-05-18T05:04:24.675Z"),
+    created: z.string().default("2026-05-18T14:04:16.993Z"),
+    updated: z.string().default("2026-05-18T14:04:16.993Z"),
     guardrail_config: z.nullable(
       z.union([
         z.lazy(() =>
@@ -3789,8 +3799,8 @@ export const CreateEvalResponseBodyFunction$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-05-18T05:04:24.675Z"),
-  updated: z.string().default("2026-05-18T05:04:24.675Z"),
+  created: z.string().default("2026-05-18T14:04:16.993Z"),
+  updated: z.string().default("2026-05-18T14:04:16.993Z"),
   guardrail_config: z.nullable(
     z.union([
       z.lazy(() =>
@@ -4008,8 +4018,8 @@ export const ResponseBodyPython$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-05-18T05:04:24.675Z"),
-  updated: z.string().default("2026-05-18T05:04:24.675Z"),
+  created: z.string().default("2026-05-18T14:04:16.993Z"),
+  updated: z.string().default("2026-05-18T14:04:16.993Z"),
   guardrail_config: z.nullable(
     z.union([
       z.lazy(() =>
@@ -4190,8 +4200,8 @@ export const ResponseBodyHTTP$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-05-18T05:04:24.675Z"),
-  updated: z.string().default("2026-05-18T05:04:24.675Z"),
+  created: z.string().default("2026-05-18T14:04:16.993Z"),
+  updated: z.string().default("2026-05-18T14:04:16.993Z"),
   guardrail_config: z.nullable(
     z.union([
       z.lazy(() =>
@@ -4367,8 +4377,8 @@ export const ResponseBodyJSON$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-05-18T05:04:24.675Z"),
-  updated: z.string().default("2026-05-18T05:04:24.675Z"),
+  created: z.string().default("2026-05-18T14:04:16.993Z"),
+  updated: z.string().default("2026-05-18T14:04:16.993Z"),
   guardrail_config: z.nullable(
     z.union([
       z.lazy(() => CreateEvalGuardrailConfigEvalsResponseBoolean$inboundSchema),
@@ -4738,8 +4748,8 @@ export const CreateEvalLlm2$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-05-18T05:04:24.675Z"),
-  updated: z.string().default("2026-05-18T05:04:24.675Z"),
+  created: z.string().default("2026-05-18T14:04:16.993Z"),
+  updated: z.string().default("2026-05-18T14:04:16.993Z"),
   guardrail_config: z.nullable(
     z.union([
       z.lazy(() =>
@@ -4760,6 +4770,7 @@ export const CreateEvalLlm2$inboundSchema: z.ZodType<
   categorical_labels: z.array(
     z.lazy(() => CreateEvalLLMEvalsCategoricalLabels$inboundSchema),
   ).optional(),
+  dataset_id: z.string().optional(),
   key: z.string(),
   mode: z.literal("jury"),
   jury: z.lazy(() => CreateEvalLLMJury$inboundSchema),
@@ -4768,6 +4779,7 @@ export const CreateEvalLlm2$inboundSchema: z.ZodType<
     "_id": "id",
     "guardrail_config": "guardrailConfig",
     "categorical_labels": "categoricalLabels",
+    "dataset_id": "datasetId",
   });
 });
 
@@ -4954,8 +4966,8 @@ export const CreateEvalLlm1$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-05-18T05:04:24.675Z"),
-  updated: z.string().default("2026-05-18T05:04:24.675Z"),
+  created: z.string().default("2026-05-18T14:04:16.993Z"),
+  updated: z.string().default("2026-05-18T14:04:16.993Z"),
   guardrail_config: z.nullable(
     z.union([
       z.lazy(() =>
@@ -4976,6 +4988,7 @@ export const CreateEvalLlm1$inboundSchema: z.ZodType<
   categorical_labels: z.array(
     z.lazy(() => CreateEvalLLMEvalsResponseCategoricalLabels$inboundSchema),
   ).optional(),
+  dataset_id: z.string().optional(),
   key: z.string(),
   mode: z.literal("single"),
   model: z.string(),
@@ -4984,6 +4997,7 @@ export const CreateEvalLlm1$inboundSchema: z.ZodType<
     "_id": "id",
     "guardrail_config": "guardrailConfig",
     "categorical_labels": "categoricalLabels",
+    "dataset_id": "datasetId",
   });
 });