+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-18T05:04:23.600Z",
+    "2026-05-18T14:04:15.855Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {