+++ packages/orq-rc/RELEASES.md
@@ -9259,3 +9259,13 @@ Based on:
 - [typescript v4.10.0-rc.11] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.11] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.11 - packages/orq-rc
+
+## 2026-05-18 14:17:52
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.1 (2.884.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.12] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.12 - packages/orq-rc